--- a/Frontend/ui-driver/src/Helpers/Storage.purs
+++ b/Frontend/ui-driver/src/Helpers/Storage.purs
@@ -128,6 +128,7 @@ data KeyStore = USER_NAME
                 | REFERRER_URL
                 | LMS_SELECTED_LANGUAGE_CACHE
                 | ANOTHER_ACTIVITY_LAUNCHED
+                | DRIVER_PROFILE_STATS_RESPONSE
 
 derive instance genericKeyStore :: Generic KeyStore _
 instance showKeyStore :: Show KeyStore where