--- a/Frontend/ui-driver/src/Screens/ChooseCityScreen/Handler.purs
+++ b/Frontend/ui-driver/src/Screens/ChooseCityScreen/Handler.purs
@@ -18,7 +18,8 @@ module Screens.ChooseCityScreen.Handler where
 
 import Control.Monad.Except.Trans (lift)
 import Control.Transformers.Back.Trans as App
-import Engineering.Helpers.BackTrack (getState)
+import Engineering.Helpers.Commons (markPerformance)
+import Engineering.Helpers.BackTrack (getState, liftFlowBT)
 import Prelude (bind, ($), pure, (<$>), discard)
 import PrestoDOM.Core.Types.Language.Flow (runScreen)
 import Screens.ChooseCityScreen.Controller (ScreenOutput(..))
@@ -30,6 +31,7 @@ import Types.ModifyScreenState (modifyScreenState)
 chooseCityScreen :: FlowBT String CHOOSE_CITY_SCREEN_OUTPUT
 chooseCityScreen = do
   (GlobalState state) <- getState
+  liftFlowBT $ markPerformance "CHOOSE_CITY_SCREEN"
   act <- lift $ lift $ runScreen $ ChooseCityScreen.screen state.chooseCityScreen
   case act of
     SelectLanguageScreen -> App.BackT $ pure App.GoBack