## Type of Change
<!-- Put an `x` in the boxes that apply -->

- [ ] Bugfix
- [ ] New feature
- [ ] Enhancement
- [ ] Refactoring
- [ ] Dependency updates

## Description
- Made All Pre-Init config fetch/updates and deep link creations parallel during Initiate.
- Moved deepLinks and viewParams from initiate to process as they were not required during initiate and were blocking.
- Parallelized Blocking RideList, DriverProfile & DriverProfileStats APIs during Initiate to obtain response from them while initiate is happening and only block for response after initiate result which ideally would not take much time considering APIs would have given back the response within the SDK initiate duration of about 2 to 3 seconds.

This brought down the App Load Time by 40% Approx locally from 9.4s to 6s.

Added Checkpoints in Code to track the time during the entire life cycle of App load to Home screen render.

| Checkpoints                        | Time          | Duration (ms) |
| ---------------------------------- | ------------- | ------------- |
|                                    | 1709854013965 |               |
| ON_CREATE_START                    | 1709854013965 | 0             |
| FORKED_INIT_TASKS_AND_APIS         | 1709854014003 | 38            |
| INIT_APP_START                     | 1709854014004 | 1             |
| INIT_APP_HYPER_SERVICE_END         | 1709854015058 | 1054          |
| INIT_HYPER_SERVICE                 | 1709854015069 | 11            |
| INIT_HYPER_SERVICE_END             | 1709854015128 | 59            |
| INDEX_BUNDLE_START                 | 1709854018974 | 3846          |
| INDEX_FIREBASE_LOG_PARAMS_START    | 1709854018978 | 4             |
| INDEX_FIREBASE_LOG_PARAMS_END      | 1709854019000 | 22            |
| INDEX_LOAD_CONFIG_START            | 1709854019001 | 1             |
| INDEX_LOAD_CONFIG_END              | 1709854019015 | 14            |
| INDEX_LOGGER_END                   | 1709854019015 | 0             |
| INDEX_BUNDLE_OS_END                | 1709854019015 | 0             |
| INDEX_BUNDLE_END_ON_MERCHANT       | 1709854019015 | 0             |
| INDEX_BUNDLE_INITIATE_START        | 1709854019020 | 5             |
| INDEX_BUNDLE_INITIATE_RESULT       | 1709854019020 | 0             |
| INIT_FUTURE_TASK_RESULT            | 1709854019022 | 2             |
| INIT_HYPER_SERVICE_INITIATE_RESULT | 1709854019030 | 8             |
| INDEX_BUNDLE_END                   | 1709854019078 | 48            |
| INDEX_PROCESS_CALLED               | 1709854019096 | 18            |
| MAIN_START                         | 1709854019097 | 1             |
| BASE_APP_FLOW_START                | 1709854019114 | 17            |
| BASE_APP_FLOW_END                  | 1709854019249 | 135           |
| INITIAL_FLOW_START                 | 1709854019250 | 1             |
| INITIAL_FLOW_END                   | 1709854019284 | 34            |
| CHECK_RIDE_AND_INITIATE_START      | 1709854019289 | 5             |
| CHECK_RIDE_AND_INITIATE_END        | 1709854019315 | 26            |
| GET_DRIVER_INFO_FLOW_START         | 1709854019315 | 0             |
| GET_DRIVER_INFO_FLOW_END           | 1709854019454 | 139           |
| HANDLE_DEEP_LINKS_FLOW             | 1709854019454 | 0             |
| CHECK_PRE_REQUISITES_FLOW          | 1709854019454 | 0             |
| CURRENT_RIDE_FLOW_START            | 1709854019490 | 36            |
| CURRENT_RIDE_FLOW_END              | 1709854019538 | 48            |
| HOME_SCREEN_FLOW                   | 1709854019538 | 0             |
| HOME_SCREEN                        | 1709854019598 | 60            |
|                        | | 5633 ms (Total Time)           |

https://docs.google.com/spreadsheets/d/1dFqRye_6wlxiB9w9EWzlVZH9tHdNM6ScI6A8Sl_TA-U/edit#gid=2036917257

### Additional Changes

- [ ] This PR modifies the database schema (database migration added)
- [ ] This PR modifies dhall configs/environment variables
- [ ] This PR contains API breaking changes
<!-- 
Provide links to the files with corresponding changes.

You can find config files in `dhall-configs`
-->


## Motivation and Context
<!--
Why is this change required? What problem does it solve?
If it fixes an open issue, please link to the issue here.

If you don't have an issue, we'd recommend starting with one first so the PR
can focus on the implementation (unless its an obvious bug or documentation fix
that will have little conversation).
-->


## How did you test it?
<!--
Did you write an integration/unit/API test to verify the code changes?
Or did you test this change manually (provide relevant screenshots)?
-->


## Checklist
<!-- Put an `x` in the boxes that apply -->

- [ ] I formatted the code and addressed linter errors `./dev/format-all-files.sh`
- [ ] I reviewed submitted code
- [ ] I added unit tests for my changes where possible
- [ ] I added a [CHANGELOG](/CHANGELOG.md) entry if applicable
- [ ] No leak detected in leakcanary